Create .gitignore
All the components that are included in the .gitignore file will not be included in the Copado Release
Management Process. Excluding unnecessary metadata will improve the performance of most of the
operations such as: Commits, validations, deployments, snapshots,...
If you need to include one of those which are already in the file, comment the line where the component is referenced.